// src/components/ui/Map/types.ts
export type Theme = "light" | "dark";